refactor(cli): Change file manipulation arguments
We change the central library to be more simple, and leave the parsing
of options to each individual application. Instead, the library will
only take `&Path` and make the necessary file operations, no need for
anything else.
Inside tasker-cli, we leave the option parsing to the execution
function, and just pass the results to the functions that will
manipulate files, cleaning the config module in the process. It's
unclear if this changes performance in any way, but it makes using the
codebase much simpler.