feat: Use IndexSet to preserve tag's insertion order
This will not be reflected immediately in all existing task files, but all newly
created tasks will see their order preserved. This clashes with the previous
behaviour in which the order the tags were printed in changed on every command
invocation. Not a real issue with the program, all things considered; mostly a
pet peeve.